Design with Theme, Display Order etc

+++ Books/Views/BookPage.xaml.cs
@@ -1,6 +1,5 @@
 using Books.Models;
-using Microsoft.Maui.Controls;
-using System.Diagnostics;
+using Books.Resources.Languages;
 
 namespace Books.Views;
 
@@ -33,6 +32,7 @@ public partial class BookPage : ContentPage
             BindingContext = note;
 
             GetImages();
+            Stars((int)note.Rating);
         }
         catch (Exception)
         {
@@ -55,6 +55,13 @@ public partial class BookPage : ContentPage
 
     async void OnDeleteButtonClicked(object sender, EventArgs e)
     {
+        bool answer = await DisplayAlert(AppResources.DeleteBookTask, AppResources.DeleteBookQuestion, AppResources.AnswerYes, AppResources.AnswerNo);
+
+        if(!answer)
+        {
+            return;
+        }
+
         var note = (Book)BindingContext;
         await App.Database.DeleteNoteAsync(note);
 
@@ -70,16 +77,6 @@ public partial class BookPage : ContentPage
         ((Stepper)sender).Value = note.Volume;
 
         VolumeLabel.Text = note.Volume.ToString();
-    }
-
-    void OnRatingStepperValueChanged(object sender, ValueChangedEventArgs e)
-    {
-        var note = (Book)BindingContext;
-        note.Rating = (float)e.NewValue;
-
-        ((Stepper)sender).Value = note.Rating;
-
-        RatingLabel.Text = note.Rating.ToString();
     }
 
     private async void PhotoButton_Clicked(object sender, EventArgs e)
@@ -105,8 +102,9 @@ public partial class BookPage : ContentPage
                 await sourceStream.CopyToAsync(localFileStream);
 
                 foldername.GalleryFolder = dir;
-                await App.Database.SaveNoteAsync(foldername);
-            }           
+            }
+
+            GetImages();
         }
         else
             await Shell.Current.DisplayAlert("Oops", "You device isn't supported", "Ok");
@@ -122,10 +120,12 @@ public partial class BookPage : ContentPage
 
             for(int i = 0; i < files.Length; i++)
             {
-                var img = new Image();
-                img.Source = files[i];                
-                img.HeightRequest = 200;
-                img.Margin = 5;
+                var img = new Image
+                {
+                    Source = files[i],
+                    HeightRequest = 200,
+                    Margin = 5
+                };
                 PhotoGrid.AddColumnDefinition(new ColumnDefinition());
                 PhotoGrid.Add(img, i, 0);
             }          
@@ -133,13 +133,168 @@ public partial class BookPage : ContentPage
         {
             //PhotoGrid.AddColumnDefinition(new ColumnDefinition());
             PhotoGrid.Add(new Label 
-                { Text = "No photos to display", 
-                FontSize = 20, 
+                { Text = AppResources.EmptyView, 
+                FontSize = 15, 
                 VerticalOptions=LayoutOptions.Center, 
                 HorizontalTextAlignment=TextAlignment.Center,
                 FontAttributes = FontAttributes.Italic,
                 TextColor = Colors.LightGrey,
                 WidthRequest=300});
         }                   
+    }
+
+    private void Fav_Clicked(object sender, EventArgs e)
+    {
+        var note = (Book)BindingContext;
+        note.Favorite = !note.Favorite;
+        Console.WriteLine("Fav ---- " + note.Favorite);
+
+        if (note.Favorite)
+            FavBtn.Source = "heart_fill.png";
+        else
+            FavBtn.Source = "heart_nofill.png";
+    }
+
+    private void Star1_Clicked(object sender, EventArgs e)
+    {
+        int star = 1;
+        var note = (Book)BindingContext;
+
+        if (note.Rating < star || note.Rating > star)
+        {
+            note.Rating = star;
+            Stars(star);
+        }
+        else
+        {
+            note.Rating = star - 1;
+            Stars(star - 1);
+        }
+    }
+
+    private void Star2_Clicked(object sender, EventArgs e)
+    {
+        int star = 2;
+        var note = (Book)BindingContext;
+
+        if (note.Rating < star || note.Rating > star)
+        {
+            note.Rating = star;
+            Stars(star);
+        }
+        else
+        {
+            note.Rating = star - 1;
+            Stars(star - 1);
+        }
+    }
+
+    private void Star3_Clicked(object sender, EventArgs e)
+    {
+        int star = 3;
+        var note = (Book)BindingContext;
+
+        if (note.Rating < star || note.Rating > star)
+        {
+            note.Rating = star;
+            Stars(star);
+        }
+        else
+        {
+            note.Rating = star - 1;
+            Stars(star - 1);
+        }
+    }
+
+    private void Star4_Clicked(object sender, EventArgs e)
+    {
+        int star = 4;
+        var note = (Book)BindingContext;
+
+        if (note.Rating < star || note.Rating > star)
+        {
+            note.Rating = star;
+            Stars(star);
+        }
+        else
+        {
+            note.Rating = star - 1;
+            Stars(star - 1);
+        }
+    }
+
+    private void Star5_Clicked(object sender, EventArgs e)
+    {
+        int star = 5;
+        var note = (Book)BindingContext;
+
+        if (note.Rating < star || note.Rating > star)
+        {
+            note.Rating = star;
+            Stars(star);
+        }
+        else
+        {
+            note.Rating = star - 1;
+            Stars(star - 1);
+        }
+    }
+
+    private void Stars(int rating)
+    {
+        string fillstar = "star_fill.png";
+        string nofillstar = "star_nofill.png";
+        
+        switch (rating)
+        {
+            case 0:
+                Console.WriteLine("CASE 0");
+                star1.Source = nofillstar;
+                star2.Source = nofillstar;
+                star3.Source = nofillstar;
+                star4.Source = nofillstar;
+                star5.Source = nofillstar;               
+                break;
+            case 1:
+                Console.WriteLine("CASE 1");
+                star1.Source = fillstar;
+                star2.Source = nofillstar;
+                star3.Source = nofillstar;
+                star4.Source = nofillstar;
+                star5.Source = nofillstar;                
+                break;
+            case 2:
+                Console.WriteLine("CASE 2");
+                star1.Source = fillstar;
+                star2.Source = fillstar;
+                star3.Source = nofillstar;
+                star4.Source = nofillstar;
+                star5.Source = nofillstar;               
+                break;               
+            case 3:
+                Console.WriteLine("CASE 3");
+                star1.Source = fillstar;
+                star2.Source = fillstar;
+                star3.Source = fillstar;
+                star4.Source = nofillstar;
+                star5.Source = nofillstar;              
+                break;
+            case 4:
+                Console.WriteLine("CASE 4");
+                star1.Source = fillstar;
+                star2.Source = fillstar;
+                star3.Source = fillstar;
+                star4.Source = fillstar;
+                star5.Source = nofillstar;              
+                break;
+            case 5:
+                Console.WriteLine("CASE 5");
+                star1.Source = fillstar;
+                star2.Source = fillstar;
+                star3.Source = fillstar;
+                star4.Source = fillstar;
+                star5.Source = fillstar;              
+                break;
+        }
     }
 }
